requirements.txt
torch
torchvision
opencv-python
pandas
matplotlib
tqdm
numpy
pillow
